+++ Changes.md
@@ -18,6 +18,7 @@ API
 ---
 
 - PathListingWidget : Added `columnContextMenuSignal()`, allowing multiple clients to collaborate on the creation of a column-specific context menu.
+- HierarchyView : Added `sceneListing()` method, to allow the context menu to be customised.
 
 Documentation
 -------------